pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 99
include = '\.pyi?$'
exclude = '''(
/(
  | \.direnv
  | \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist

)/
)
'''

[tool.isort]
profile = "black"
line_length = 99
known_first_party = ["movici_api_client"]

[tool.poetry]
name = "movici-api-client"
version = "0.1.0"
description = "Client library and CLI tool for interfacing with the Movici Cloud Platform"
authors = ["Pelle Koster <pelle.koster@nginfra.nl>"]
license = "Movici Public License"
readme = "README.rst"
packages = [{ include = "movici_api_client", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
httpx = "^0.23.1"
click = "^8.1.3"
questionary = "^1.10.0"
tabulate = "^0.9.0"
tqdm = "^4.64.1"
gimme-that = "^0.3.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
flake8 = "^6.0.0"
black = "^22.12.0"
isort = "^5.11.2"
pytest-asyncio = "^0.20.3"

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.poetry.scripts]
movici = 'movici_api_client.cli:__main__'